Sample Application - PHP & MySQL
This is a simple application that allows users to manage a list of people. It contains generic functions for adding, editing, and deleting records from a MySQL DB via a PHP web application. The files are presented here to illustrate separation of functions using an MVC approach. 

The HUMANS.sql file contains sample data for testing.